// src/types/League.ts
export interface League {
    id: string;
    name: string;
    shortName: string;
    description: string;
    logo: string;
    formedYear: number;
    badge: string;
    banner: string;
    facebook: string;
    website: string;
    x: string;
    youtube: string;
    sport: string;
    rss: string;
    instagram: string;
    poster: string;
    fanArt: string[];
    trophy: string;
    firstEventDate: string;
}