// src/types/Event.ts
export interface Event {
    idEvent: string;
    idLeague: string;
    idVenue: string;
    strEvent: string;
    strPoster: string;
    intRound: number;
    strBanner: string;
    strMap?: string;
    strLeagueBadge: string;
    strVenue: string;
    strCity: string;
    strCountry: string;
    strLeague: string;
    strPostponed: string;
    strSeason: string;
    strThumb: string;
    strTime: string;
    strTimeLocal: string;
    strTimestamp: string;
}